Refactor ML training pipeline, artifact diagnostics, and production model resolution
- Hard-code calendar-based train/validation/test windows in LSTM, Random Forest, and XGBoost training scripts with editable top-level constants.
- Default training outputs to project models directory and remove CLI split/date controls now superseded by fixed calendar constants.
- Extend global classification dataset builder to support explicit date-range splitting with split-strategy metadata.
- Upgrade XGBoost training robustness: add parallel n_jobs wiring, adaptive threshold candidate grid, validation prediction-diversity diagnostics, and median-threshold fallback before failing strict anti-degeneracy checks.
- Preserve GPU/CPU fallback behavior in script-level XGBoost runtime configuration and persist effective GPU usage in run metadata.
- Add n_jobs support to RandomForest model constructor and mixed-precision-safe float32 output head for LSTM model.
- Introduce shared production model config utilities (normalize/load/resolve classifier type + compatible model-name aliases).
- Update UI chat candidate selection to prefer production-config model names, with fallback to all predictions when filtered candidates are unavailable.
- Enhance classifier and LSTM visualizations with artifact freshness diagnostics versus latest market date and add classifier health warnings for near-degenerate behavior.
- Surface stale-artifact and classifier-health warnings in Streamlit views for faster operator feedback.

+++ src/models/lstm_model.py
@@ -33,7 +33,7 @@ class LSTMModel(StockPredictor):
                 Input(shape=(self.sequence_length, self.n_features)),
                 LSTM(128, return_sequences=True),
                 LSTM(64, return_sequences=False),
-                Dense(1),
+                Dense(1, dtype="float32"),
             ]
         )
